chore: try `/edge` entrypoint
